feature/DBCON-298: moves OracleCredentialsMaskUtils class to oracle.util package
